+++ src/zope/component/interfaces.py
@@ -364,6 +364,12 @@ class IFactory(Interface):
 
 def _inherits_docs(func, iface):
     doc = iface[func.__name__].__doc__
+
+    # doc can be None if the interpreter drops all docstrings when the
+    # environment variable PYTHONOPTIMIZE is set 2.
+    if not doc:
+        return func
+
     # By adding the ..seealso:: we get a link from our overview page
     # to the specific narrative place where the function is described, because
     # our overview page uses :noindex: